requirements.txt
Django==1.8.3
MySQL-python
django-crispy-forms==1.4.0
